refactor(ai-gateway): inline Novita, Ramp, Neon; fix Ramp bare model ids
